async DB connection, models moved to one file!
