+++ app/conf/config.py
@@ -35,7 +35,9 @@ class Settings(BaseSettings):
     cloudinary_api_secret: str = 'secret'
 
     def postgres_url(self) -> str:
-        return f"postgresql://{self.postgres_user}:{self.postgres_password}@{self.postgres_host}:{self.postgres_port}"
+        # return f"postgresql://{self.postgres_user}:{self.postgres_password}@{self.postgres_host}:{self.postgres_port}"
+        return f"postgresql+asyncpg://{self.postgres_user}:{self.postgres_password}@{self.postgres_host}:{self.postgres_port}"
+
 
 
 settings = Settings()
