--- app/database/post.py
@@ -1,33 +1,0 @@
-from sqlalchemy import (
-    Column,
-    Integer,
-    String,
-    DateTime,
-    ForeignKey,
-    Table,
-    func
-)
-from sqlalchemy.orm import relationship
-
-from app.database.db import Base
-
-post_m2m_tag = Table(
-    "post_m2m_tag",
-    Base.metadata,
-    Column("id", Integer, primary_key=True),
-    Column("post_id", Integer, ForeignKey("posts.id", ondelete="CASCADE")),
-    Column("tag_id", Integer, ForeignKey("tags.id", ondelete="CASCADE"))
-)
-
-
-class Post(Base):
-    __tablename__ = "posts"
-    id = Column(Integer, primary_key=True, index=True)
-    photo = Column(String(255), nullable=False)
-    description = Column(String(255), nullable=False)
-    created_at = Column('created_at', DateTime, default=func.now())
-    updated_at = Column("updated_at", DateTime, default=func.now(), onupdate=func.now())
-    comments = relationship("Comment", back_populates="post")
-    tags = relationship("Tag", secondary=post_m2m_tag, backref="posts")
-    user_id = Column('user_id', ForeignKey('users.id', ondelete='CASCADE'), default=None)
-    user = relationship('User', backref="posts")
