--- app/database/user.py
@@ -1,16 +1,0 @@
-from sqlalchemy import Column, Integer, String, func, Boolean
-from sqlalchemy.sql.sqltypes import DateTime
-from app.database.db import Base
-
-
-class User(Base):
-    __tablename__ = "users"
-    id = Column(Integer, primary_key=True)
-    email = Column(String(250), nullable=False, unique=True)
-    first_name = Column(String(250), nullable=False)
-    last_name = Column(String(250), nullable=False)
-    password = Column(String(255), nullable=False)
-    created_at = Column(DateTime, default=func.now())
-    avatar = Column(String(255), nullable=True)
-    refresh_token = Column(String(255), nullable=True)
-    confirmed = Column(Boolean, default=False)
