+++ pyproject.toml
@@ -26,6 +26,7 @@ fastapi-mail = "^1.4.1"
 python-jose = "^3.3.0"
 bcrypt = "^4.1.3"
 cloudinary = "^1.40.0"
+asyncpg = "^0.29.0"
 
 [tool.poetry.scripts]
 start = "uvicorn main:app --host localhost --port 8008 --reload"
